refactor(JwtUtil): auth 하위 디렉토리로 위치 이동
