+++ src/main/java/com/chaeum/api/global/filter/CustomEntryPoint.java
@@ -3,6 +3,7 @@ package com.chaeum.api.global.filter;
 import com.chaeum.api.global.exception.ErrorCode;
 import com.chaeum.api.global.response.ApiResponse;
 import com.chaeum.api.global.response.ErrorResponse;
+import com.chaeum.api.global.utils.ResponseUtil;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -19,23 +20,13 @@ import java.io.IOException;
 @Component
 public class CustomEntryPoint implements AuthenticationEntryPoint {
 
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
     @Override
     public void commence(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            AuthenticationException e
-    ) throws IOException, ServletException {
-
-        log.error("Unauthorized access attempt: {}", e.getMessage());
-        createAPIResponse(response, ErrorCode.MEMBER_NOT_AUTHENTICATED);
-    }
-
-    private void createAPIResponse(HttpServletResponse response, ErrorCode errorCode) throws IOException {
-        response.setStatus(errorCode.getStatus().value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-        objectMapper.writeValue(response.getWriter(), ErrorResponse.error(errorCode));
+        HttpServletRequest request,
+        HttpServletResponse response,
+        AuthenticationException e
+    ) throws IOException {
+        log.warn("Unauthorized access to {}: {}", request.getRequestURI(), e.getMessage());
+        ResponseUtil.writeError(response, ErrorCode.MEMBER_NOT_AUTHENTICATED);
     }
 }
