+++ src/main/java/com/chaeum/api/global/filter/CustomLogoutFilter.java
@@ -3,10 +3,8 @@ package com.chaeum.api.global.filter;
 import com.chaeum.api.global.auth.repository.RefreshTokenRepository;
 import com.chaeum.api.global.exception.ChaeumException;
 import com.chaeum.api.global.exception.ErrorCode;
-import com.chaeum.api.global.response.ApiResponse;
-import com.chaeum.api.global.response.ErrorResponse;
-import com.chaeum.api.global.utils.JwtUtil;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.chaeum.api.global.auth.util.JwtUtil;
+import com.chaeum.api.global.utils.ResponseUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
@@ -14,105 +12,84 @@ import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.web.filter.GenericFilterBean;
-
 import java.io.IOException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.filter.GenericFilterBean;
 
 @RequiredArgsConstructor
 public class CustomLogoutFilter extends GenericFilterBean {
+
+    private static final String REFRESH_COOKIE_NAME = "refresh";
 
     private final JwtUtil jwtUtil;
     private final RefreshTokenRepository refreshTokenRepository;
 
     @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain)
-            throws IOException, ServletException {
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+        throws IOException, ServletException {
+
         HttpServletRequest httpRequest = (HttpServletRequest) request;
         HttpServletResponse httpResponse = (HttpServletResponse) response;
 
-        // 1. 로그아웃 요청인지 확인 (아니면 다음 필터로 넘김)
         if (!isLogoutRequest(httpRequest)) {
-            filterChain.doFilter(request, response);
+            chain.doFilter(request, response);
             return;
         }
 
-        // 2. Refresh Token 쿠키에서 추출 (없으면 예외 발생)
         String refreshToken = extractRefreshToken(httpRequest);
-
-        // 3. Refresh Token 유효성 검증 (문제 있으면 예외 응답 반환)
         if (!validateRefreshToken(refreshToken, httpResponse)) {
             return;
         }
 
-        // 4. Redis에서 Refresh Token 삭제
         refreshTokenRepository.deleteById(jwtUtil.getEmail(refreshToken));
-
-        // 5. Refresh Token 쿠키 제거 후 응답 반환
         removeRefreshTokenCookie(httpResponse);
-        sendSuccessApiResponse(httpResponse);
+        ResponseUtil.writeSuccess(httpResponse);
     }
 
-    // 로그아웃 확인
     private boolean isLogoutRequest(HttpServletRequest request) {
-        return request.getRequestURI().equals("/api/v1/logout") && "POST".equals(request.getMethod());
+        return "/api/v1/logout".equals(request.getRequestURI()) && "POST".equalsIgnoreCase(request.getMethod());
     }
 
-    // 쿠키에서 Refresh Token 추출
     private String extractRefreshToken(HttpServletRequest request) {
         if (request.getCookies() == null) {
             throw ChaeumException.from(ErrorCode.NOT_FOUND_REFRESH_TOKEN);
         }
         for (Cookie cookie : request.getCookies()) {
-            if ("refresh".equals(cookie.getName())) {
+            if (REFRESH_COOKIE_NAME.equals(cookie.getName())) {
                 return cookie.getValue();
             }
         }
         throw ChaeumException.from(ErrorCode.NOT_FOUND_REFRESH_TOKEN);
     }
 
-    // Refresh Token 검증(유효하면 true)
-    private boolean validateRefreshToken(String refreshToken, HttpServletResponse response) throws IOException {
-        if (!jwtUtil.validateToken(refreshToken)) {
-            sendErrorApiResponse(response, ErrorCode.INVALID_REFRESH_TOKEN);
+    private boolean validateRefreshToken(String token, HttpServletResponse response) throws IOException {
+        if (!jwtUtil.validateToken(token)) {
+            ResponseUtil.writeError(response, ErrorCode.INVALID_REFRESH_TOKEN);
             return false;
         }
-        if (jwtUtil.isExpired(refreshToken)) {
-            sendErrorApiResponse(response, ErrorCode.EXPIRED_REFRESH_TOKEN);
+
+        if (jwtUtil.isExpired(token)) {
+            ResponseUtil.writeError(response, ErrorCode.EXPIRED_REFRESH_TOKEN);
             return false;
         }
-        if (!"refresh".equals(jwtUtil.getCategory(refreshToken))) {
-            sendErrorApiResponse(response, ErrorCode.INVALID_REFRESH_TOKEN);
+
+        if (!"refresh".equals(jwtUtil.getCategory(token))) {
+            ResponseUtil.writeError(response, ErrorCode.INVALID_REFRESH_TOKEN);
             return false;
         }
-        if (refreshTokenRepository.findById(jwtUtil.getEmail(refreshToken)).isEmpty()) {
-            sendErrorApiResponse(response, ErrorCode.INVALID_REFRESH_TOKEN);
+
+        if (refreshTokenRepository.findById(jwtUtil.getEmail(token)).isEmpty()) {
+            ResponseUtil.writeError(response, ErrorCode.INVALID_REFRESH_TOKEN);
             return false;
         }
+
         return true;
     }
 
-    // Refresh Token 쿠키 삭제
     private void removeRefreshTokenCookie(HttpServletResponse response) {
-        Cookie cookie = new Cookie("refresh", null);
+        Cookie cookie = new Cookie(REFRESH_COOKIE_NAME, null);
         cookie.setMaxAge(0);
         cookie.setPath("/");
         response.addCookie(cookie);
-    }
-
-    private void sendSuccessApiResponse(HttpServletResponse response) throws IOException {
-        response.setStatus(HttpStatus.OK.value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-        new ObjectMapper().writeValue(response.getWriter(), ApiResponse.success());
-    }
-
-    private void sendErrorApiResponse(HttpServletResponse response, ErrorCode errorCode) throws IOException {
-        response.setStatus(errorCode.getStatus().value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-        new ObjectMapper().writeValue(response.getWriter(), ErrorResponse.error(errorCode));
     }
 }
